pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mic-hackathon-2025"
version = "0.1.0"
description = "A short description of the research project"
readme = "README.md"
requires-python = ">= 3.12"
license = { file = "LICENSE" }
authors = [{ name = "Ankit Shrivastava" }]
keywords = [""]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
]

dependencies = [
  "numpy",
  "matplotlib",
  "jupyterlab",
  "ruamel.yaml",
  "torchvision",
  "torchinfo",
  "tensorboard",
  "h5py"
]

[project.optional-dependencies]
docs = [
  "sphinx",
  "sphinx-rtd-theme",
  "nbsphinx",
  "myst-parser",
  "sphinx-last-updated-by-git",
  "sphinx-autodoc-typehints",
]
dev = [
  "pytest",
  "pytest-cov",

  # --- Linting & Formatting ---
  "black==24.10.0",
  "ruff==0.7.0",
  "mypy==1.13.0",
  "pre-commit",

  # --- Debugging / REPL / Notebook ---
  "ipython==8.29.0",
  "ipykernel==6.29.5",
  "ipdb==0.13.13",
  "debugpy==1.8.7",

  # --- Profilers ---
  "line_profiler==4.1.3",
  "memory_profiler",
  "scalene==1.5.33.1",
]


# -------------------------
# mypy
# -------------------------
[tool.mypy]
files = ["experiment/codes/scripts", "experiment/codes/utils"]
mypy_path = ["experiment/codes", "experiment/codes/scripts"]
python_version = "3.12"
ignore_missing_imports = false
explicit_package_bases = true
follow_imports = "normal"
warn_unused_configs = true
warn_redundant_casts = false
warn_unused_ignores = false
disallow_untyped_defs = false
check_untyped_defs = false
no_implicit_optional = false
exclude = '(\.venv|venv|env|\.env|\.mypy_cache|build|dist|_build|data/raw|data/processed)'

[[tool.mypy.overrides]]
module = ["ruamel.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["torchvision.*"]
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["torch.*"]
ignore_missing_imports = true
follow_imports = "skip"

# -------------------------
# black
# -------------------------
[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'

exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.pytest_cache
  | \.ipynb_checkpoints
  | __pycache__
  | \.venv
  | venv
  | env
  | \.env
  | \.virtualenv
  | \.virtualenvs
  | build
  | dist
  | _build
  | outputs?
  | checkpoints?
  | data/raw
  | data/processed
)/
'''

# -------------------------
# ruff
# -------------------------
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = [
  "E203", # Black conflict
]

exclude = [
  ".git",
  ".hg",
  ".mypy_cache",
  ".pytest_cache",
  ".ipynb_checkpoints",
  "__pycache__",
  ".venv",
  "venv",
  "env",
  ".env",
  ".virtualenv",
  ".virtualenvs",
  "_build",
  "build",
  "dist",
  "data/raw",
  "data/processed",
]

# -------------------------
# pytest
# -------------------------
[tool.pytest.ini_options]
pythonpath = ["."]
minversion = "6.0"
addopts = [
  "-v",
  "-ra",
  "--strict-markers",
  "--disable-warnings",
  # "--cov=src",
  # "--cov-report=term-missing",
]

testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
markers = [
  "slow: slow tests",
  "integration: integration tests",
]

# -------------------------
# Coverage (pytest-cov)
# -------------------------
# [tool.coverage.run]
# branch = true
# source = ["src"]

# [tool.coverage.report]
# show_missing = true
# skip_covered = true
# precision = 1

# -------------------------
# setuptools: package discovery
# -------------------------

# [tool.setuptools]
# package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*", "utils", "utils.*"]
